refactor: adopt @ alias for frontend imports
