+++ client/src/pages/MyBookingsPage.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from "react";
 import { Link } from "react-router-dom";
-import { fairroomApi } from "../api/fairroomApi";
-import type { BookingSummary, Reminder, Room } from "../api/contracts";
+import { fairroomApi } from "@/api/fairroomApi";
+import type { BookingSummary, Reminder, Room } from "@/api/contracts";
 
 const timeLabel = (value: string) =>
   new Intl.DateTimeFormat("en-GB", {
