+++ client/src/pages/RoomDetailsPage.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from "react";
 import { Link, useLocation, useParams } from "react-router-dom";
-import { fairroomApi } from "../api/fairroomApi";
-import type { AvailabilityWindow, Room } from "../api/contracts";
+import { fairroomApi } from "@/api/fairroomApi";
+import type { AvailabilityWindow, Room } from "@/api/contracts";
 
 const toDateTimeLabel = (value: string) =>
   new Intl.DateTimeFormat("en-GB", {
