+++ client/src/types/types.ts
@@ -1,3 +1,3 @@
 // Legacy barrel kept for any leftover imports.
 // Canonical frontend types live in ../api/contracts.
-export type { Room, UserProfile as User, UserRole } from "../api/contracts";
+export type { Room, UserProfile as User, UserRole } from "@/api/contracts";
